pyproject.toml
[tool.poetry]
name = "geolocation-determinant"
version = "0.1.0"
description = ""
authors = ["bulat <bulatsubuh@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.5"}
apscheduler = "^3.10.4"
motor = "^3.6.0"
pydantic = "^2.9.2"
asyncio = "^3.4.3"
pandas = "^2.2.3"
geocoder = "^1.38.1"
gitpython = "^3.1.43"
numpy = "^2.1.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
